% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codexdata-package.r
\docType{package}
\name{codexdata-package}
\alias{codexdata}
\alias{codexdata-package}
\title{Tidy Codex game data and summarise model results}
\description{
Match data for tournament mathces of Codex by Sirlin Games.
}
\author{
\strong{Maintainer}: Mark Webster \email{markwebster204@yahoo.co.uk}

}
\keyword{internal}
