% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nicknames}
\alias{nicknames}
\title{Common Codex deck nicknames}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 12 rows and 2 columns.}
\usage{
nicknames
}
\description{
Common Codex deck nicknames
}
\keyword{datasets}
