% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_deck_names.r
\name{prepare_deck_names_for_modelling}
\alias{prepare_deck_names_for_modelling}
\title{Extract specs and starters from deck (nick)names for modelling}
\usage{
prepare_deck_names_for_modelling(
  name,
  starters,
  nicknames = data.table(name = character(0), nickname = character(0))
)
}
\arguments{
\item{name}{a character vector, containing deck names in standardised form, or their nicknames}

\item{starters}{a data.table, containing columns for spec names and their associated starter colour}

\item{nicknames}{an optional data.table, containing standardised deck names
and their associated nicknames, so deck nicknames can be accepted as input}
}
\value{
a tibble, containing a column for the deck's starter name and three columns for its spec names.
}
\description{
Extract specs and starters from deck (nick)names for modelling
}
