% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{starters}
\alias{starters}
\title{Specification-starter lookup table}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 20 rows and 3 columns.}
\usage{
starters
}
\description{
Specification-starter lookup table
}
\keyword{datasets}
