% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_deck_names.r
\name{standardise_deck_name}
\alias{standardise_deck_name}
\title{Standardise recorded deck names, change to any common nicknames}
\usage{
standardise_deck_name(
  name,
  starters = starters,
  nicknames = data.table(name = character(0), nickname = character(0)),
  return_nicknames = TRUE
)
}
\arguments{
\item{name}{a character vector, containing deck names or nicknames}

\item{starters}{a data.table, containing columns for spec names and their associated starter colour}

\item{nicknames}{an optional data.table, containing standardised deck names
and their associated nicknames, so deck nicknames can be accepted as input}

\item{return_nicknames}{if TRUE, returns a deck's nickname instead of its name if there is one in \code{nicknames}}
}
\value{
a character vector of the same length as \code{name}, containing
  standardised versions of the deck names, or their nicknames when there is
  one.
}
\description{
Standardise recorded deck names, change to any common nicknames
}
